# Speedup merging of parse error