I don't want to use full blown Dictionary since most of the time we have only one item in it. So shortcircuit things for simple cases make code faster.

## Before

| Method                 | Mean           | Error         | StdDev        | Median         |
|----------------------- |---------------:|--------------:|--------------:|---------------:|
| ExpressionParser       | 6,213,672.3 us | 201,067.50 us | 592,852.12 us | 5,941,961.0 us |
| ManualExpressionParser |    18,644.7 us |     263.30 us |     246.29 us |    18,598.4 us |
| Lex                    |       946.6 us |       5.51 us |       5.16 us |       947.2 us |

## After

| Method                 | Mean           | Error        | StdDev      |
|----------------------- |---------------:|-------------:|------------:|
| ExpressionParser       | 3,326,861.3 us | 12,140.53 us | 9,478.53 us |
| ManualExpressionParser |    13,073.6 us |    111.79 us |   104.56 us |
| Lex                    |       945.7 us |      9.11 us |     8.52 us |